+++ core/src/basic/mapx_raw/mod.rs
@@ -389,6 +389,13 @@ impl MapxRaw {
         self.inner.as_prefix_slice()
     }
 
+    /// Returns the unique instance ID of this `MapxRaw`.
+    pub fn instance_id(&self) -> u64 {
+        let mut bytes = [0u8; 8];
+        bytes.copy_from_slice(self.as_prefix_slice());
+        u64::from_be_bytes(bytes)
+    }
+
     /// Checks if this `MapxRaw` instance is the same as another.
     ///
     /// # Arguments